requirements.txt
appnope             0.1.4
asttokens           2.4.1
branca              0.8.0
certifi             2024.8.30
charset-normalizer  3.3.2
click               8.1.7
comm                0.2.2
configparser        7.1.0
contourpy           1.3.0
cycler              0.12.1
debugpy             1.8.6
decorator           5.1.1
exceptiongroup      1.2.2
executing           2.1.0
folium              0.17.0
fonttools           4.54.1
geographiclib       2.0
geopandas           1.0.1
geopy               2.4.1
idna                3.10
importlib_metadata  8.5.0
importlib_resources 6.4.5
ipykernel           6.29.5
ipython             8.18.1
jedi                0.19.1
Jinja2              3.1.4
joblib              1.4.2
jupyter_client      8.6.3
jupyter_core        5.7.2
kiwisolver          1.4.7
MarkupSafe          3.0.1
matplotlib          3.9.2
matplotlib-inline   0.1.7
nest-asyncio        1.6.0
nltk                3.9.1
numpy               2.0.2
packaging           24.1
pandas              2.2.3
parso               0.8.4
pexpect             4.9.0
pillow              10.4.0
pip                 21.2.4
platformdirs        4.3.6
prompt_toolkit      3.0.48
psutil              6.0.0
ptyprocess          0.7.0
pure_eval           0.2.3
Pygments            2.18.0
pyogrio             0.10.0
pyparsing           3.1.4
pyproj              3.6.1
python-dateutil     2.9.0.post0
python-dotenv       1.0.1
pytz                2024.2
pyzmq               26.2.0
regex               2024.9.11
requests            2.32.3
scikit-learn        1.5.2
scipy               1.13.1
seaborn             0.13.2
setuptools          58.0.4
shapely             2.0.6
six                 1.16.0
sklearn             0.0
stack-data          0.6.3
threadpoolctl       3.5.0
tornado             6.4.1
tqdm                4.66.6
traitlets           5.14.3
typing_extensions   4.12.2
tzdata              2024.2
urllib3             2.2.3
wcwidth             0.2.13
xyzservices         2024.9.0
zipp                3.20.2
